WIP Phase 3.
The trolls are there, but right now they do not kill the player. The AI is also incredibly slow for large paths.

Re-did the tiling system so that an entity can define a tile for each direction rather than having the component set this stuff.

The game is now truly turn-based. Gone is the timer, the game now blocks on input and does not do anything until input is received.
There is an early exit thing right after the input handler call to allow the application to quit.

So TODO:
- Make trolls kill the player
- Make troll AI faster
- Trolls/AI should not move when turning, just like the player (i.e. spend a turn). Otherwise they have an unfair advantage.
- Make level end on X.