Fixed crushed troll tile rendering.
The player physics would not always run first, so trolls would not always know whether they had been crushed by the player. I've now added
a new nominated pass called priorityPhysics, which runs before the general physics pass. This should be used by the player entity since the
player always starts the turn. Fixed a bug with the text renderer that would allow negative x-coordinates to be used for RenderMessages.

Also increased the number of trolls to 20 and reduced their path lengths to 100 for a greater challenge.